Move new tests to parallel
